Create tab selector in left sidebar
